Extract GetAccounts, RemoveAsync into PCA and CCA from ClientAppBase. Refactor ServiceBundle into CCA/PCA versions.
