+++ src/client/Microsoft.Identity.Client.Services/Internal/IServiceBundle.cs
@@ -7,7 +7,6 @@ using Microsoft.Identity.Client.Instance.Discovery;
 using Microsoft.Identity.Client.OAuth2.Throttling;
 using Microsoft.Identity.Client.PlatformsCommon.Interfaces;
 using Microsoft.Identity.Client.TelemetryCore;
-using Microsoft.Identity.Client.WsTrust;
 
 namespace Microsoft.Identity.Client.Internal
 {
@@ -23,7 +22,6 @@ namespace Microsoft.Identity.Client.Internal
         IHttpManager HttpManager { get; }
         IInstanceDiscoveryManager InstanceDiscoveryManager { get; }
         IPlatformProxy PlatformProxy { get; }
-        IWsTrustWebRequestManager WsTrustWebRequestManager { get; }
         IDeviceAuthManager DeviceAuthManager { get; }
         IThrottlingProvider ThrottlingManager { get; }
